feat: * remove last remnant of old cache * remove mini-moka dependency * implement once cache for creature fields function * update dependencies
