+++ Cargo.toml
@@ -31,14 +31,13 @@ utoipa = { version = "4.2.3", features = ["actix_extras"] }
 utoipa-swagger-ui = { version = "7.0.1", features = ["actix-web"] }
 
 sqlx = { version = "0.7.4", features = ["runtime-async-std", "sqlite"] }
-mini-moka = "0.10.3"
 cached = { version = "0.51.3", features = ["async"] }
 
-anyhow = "1.0.83"
-serde = { version = "1.0.201", features = ["derive"] }
+anyhow = "1.0.86"
+serde = { version = "1.0.202", features = ["derive"] }
 serde_json = "1.0.117"
 strum = {version="0.26.2", features = ["derive"]}
-rand = "0.9.0-alpha.0"
+rand = "0.9.0-alpha.1"
 counter = "0.5.7"
 dotenvy = "0.15.7"
 regex = "1.10.4"
@@ -50,6 +49,6 @@ num-traits = "0.2.19"
 
 [build-dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "rt"] }
-anyhow = "1.0.83"
+anyhow = "1.0.86"
 sqlx = {version = "0.7.4", features = ["runtime-async-std", "sqlite"]}
 dotenvy = "0.15.7"
