+++ src/db/mod.rs
@@ -1,4 +1,3 @@
-pub mod cache;
 pub mod cr_core_initializer;
 pub mod data_providers;
 pub mod proxy;
