+++ src/models/creature_component/fields_unique_values_struct.rs
@@ -1,0 +1,11 @@
+#[derive(Default, Eq, PartialEq, Clone)]
+pub struct FieldsUniqueValuesStruct {
+    pub list_of_levels: Vec<String>,
+    pub list_of_families: Vec<String>,
+    pub list_of_traits: Vec<String>,
+    pub list_of_sources: Vec<String>,
+    pub list_of_alignments: Vec<String>,
+    pub list_of_sizes: Vec<String>,
+    pub list_of_rarities: Vec<String>,
+    pub list_of_creature_types: Vec<String>,
+}
