+++ src/models/creature_component/mod.rs
@@ -3,4 +3,5 @@ pub mod creature_core;
 pub mod creature_extra;
 pub mod creature_spell_caster;
 pub mod creature_variant;
+pub mod fields_unique_values_struct;
 pub mod filter_struct;
